refactor: 整合ConsulRegistry
