+++ motan-registry-consul/src/main/resources/META-INF/services/com.weibo.api.motan.registry.RegistryFactory
@@ -15,4 +15,3 @@
 #
 
 com.weibo.api.motan.registry.consul.ConsulRegistryFactory
-com.weibo.api.motan.registry.consul.command.CommandConsulRegistryFactory
